// src/usuario/dto/usuario.dto.ts
export class UsuarioDto {

    primer_nombre: string;
    segundo_nombre: string;
    primer_apellido: string;
    segundo_apellido: string;
    username: string;
    fecha_nacimiento: string;
    genero: string;
    email: string;
    mobile: string;
    clave: string;
    rol: string;
    //direcciones de entrega

}